% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_mjsplot.R
\name{as_mjsplot}
\alias{as_mjsplot}
\title{Turn a simple ggplot plot into an mjs_plot object}
\usage{
as_mjsplot(gg)
}
\arguments{
\item{gg}{ggplot object}
}
\value{
metricsgraphics object
}
\description{
Takes a ggplot object that has a single geom (it can be geom_line,
geom_point or geom_histogram) and converts it to it's metricsgraphics counterpart.
It will do it's best to identify plot labels and, in the case of
geom_histogram, the number of bins.
}
\examples{
dat <- data.frame(year=seq(1790, 1970, 10),
                  uspop=as.numeric(uspop))

set.seed(5689)
movies <- movies[sample(nrow(movies), 1000), ]

gg1 <- ggplot(dat, aes(x=year, y=uspop)) + geom_line()
as_mjsplot(gg1)

gg2 <- ggplot(dat, aes(x=year, y=uspop)) + geom_point()
as_mjsplot(gg2)

gg3 <- ggplot(movies, aes(rating)) + geom_histogram()
as_mjsplot(gg3)

gg4 <- ggplot(movies, aes(rating)) + geom_histogram(binwidth = 0.1)
as_mjsplot(gg4)
}

