% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metricsgraphics-package.R
\docType{package}
\name{metricsgraphics}
\alias{metricsgraphics}
\alias{metricsgraphics-package}
\title{An \code{htmlwidget} interface to the
\href{http://metricsgraphicsjs.org/}{MetricsGraphics.js} D3 chart library}
\description{
An \code{htmlwidget} interface to the
\href{http://metricsgraphicsjs.org/}{MetricsGraphics.js} D3 chart library
}
\author{
Bob Rudis (@hrbrmstr)
}

