% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_add_mouseover}
\alias{mjs_add_mouseover}
\title{Adds a custom rollover to a metricsgraphics chart}
\usage{
mjs_add_mouseover(mjs, func)
}
\arguments{
\item{mjs}{plot object}

\item{func}{text for javascript function to be used for the custom rollover. See Details for usage.}
}
\value{
metricsgraphics object
}
\description{
Adds a custom rollover to a metricsgraphics chart
}
\details{
MetricsGraphics charts allow for \href{https://github.com/mozilla/metrics-graphics/wiki/Graphic#mouseover}{custom rollovers}.
    \code{mjs_add_mouseover} lets you add a custom rollover to a metricsgraphics object. You must be
    familiar with javascript and D3 idioms since you are supplying a javascript function as
    a parameter.\cr
    \cr
    Since targeting is done by element id, you will need to add a special string - \code{\{\{ID\}\}} -
    to the target element selector so metricsgraphics can add the unique object identifier
    to the selector. See Examples for basic usage.
}
\examples{
\dontrun{
dat \%>\%
  mjs_plot(x=date, y=value) \%>\%
  mjs_line() \%>\%
  mjs_axis_x(xax_format = "date") \%>\%
  mjs_add_mouseover("function(d, i) {
                $('{{ID}} svg .mg-active-datapoint')
                    .text('custom text : ' + d.date + ' ' + i);
                 }")
}
}

