% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_axis_y}
\alias{mjs_axis_y}
\title{Configure y axis ticks & limits}
\usage{
mjs_axis_y(mjs, yax_count = 5, min_y = NULL, max_y = NULL,
  y_scale_type = "linear")
}
\arguments{
\item{mjs}{plot object}

\item{yax_count}{tick count}

\item{min_y}{min limit for y axis}

\item{max_y}{max limit for y axis}

\item{y_scale_type}{scale for y axis; either "linear" (default) or "log"}
}
\description{
Configure y axis ticks & limits
}

