% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_grid}
\alias{mjs_grid}
\title{Draw separate charts for each line}
\usage{
mjs_grid(mjs, grid = TRUE)
}
\arguments{
\item{grid}{if TRUE, draw separate charts for each line}
}
\description{
Draw separate charts for each line
}
\examples{
\dontrun{
set.seed(1492)
stocks <- data.frame(
  time = as.Date('2009-01-01') + 0:9,
  X = rnorm(10, 0, 1),
  Y = rnorm(10, 0, 2),
  Z = rnorm(10, 0, 4))

stocks \%>\%
  mjs_plot(x=time, y=X) \%>\%
  mjs_line() \%>\%
  mjs_add_line(Y) \%>\%
  mjs_add_line(Z) \%>\%
  mjs_axis_x(xax_format="date") \%>\%
  mjs_add_legend(legend=c("X", "Y", "Z")) \%>\%
  mjs_grid()
}
}

