% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_labs}
\alias{mjs_labs}
\title{Configure axis labels & plot description}
\usage{
mjs_labs(mjs, x_label = NULL, y_label = NULL)
}
\description{
Configure axis labels & plot description
}
\examples{
\dontrun{
mtcars \%>\%
 mjs_plot(x=wt, y=mpg, width=400, height=300) \%>\%
 mjs_point(color_accessor=carb, size_accessor=carb) \%>\%
 mjs_labs(x="Weight of Car", y="Miles per Gallon")
}
}

