% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as_mjsplot.R
\name{mjs_plot_}
\alias{mjs_plot_}
\title{Create a new metricsgraphics.js  plot}
\usage{
mjs_plot_(data, x, y, show_rollover_text = TRUE, linked = FALSE,
  decimals = 2, format = "count", missing_is_hidden = FALSE, left = 80,
  right = 10, top = 40, bottom = 60, buffer = 8, width = NULL,
  height = NULL)
}
\arguments{
\item{data}{data frame}

\item{x}{variable holding character name for data.frame column}

\item{y}{varaible holding character name for data.frame column}

\item{show_rollover_text}{determines whether or not to show any text when a data point is rolled over.}

\item{linked}{inks together all other graphs whose linked option is set to true.
When one graphic in that set is rolled over, the corresponding values in the other
graphics are also rolled over (default: \code{FALSE} - not linked)}

\item{decimals}{the number of decimals to show in a rollover (default: \code{2})}

\item{format}{sets the format of the data object, which is to say,
\code{count}s or \code{percentage}s}

\item{missing_is_hidden}{if true and if the data object is a time series, missing data points will be treated as zeros}

\item{left}{the size of the left margin in pixels.}

\item{right}{the size of the right margin in pixels.}

\item{top}{the size of the top margin in pixels.}

\item{bottom}{the size of the bottom margin in pixels.}

\item{buffer}{the buffer size in pixels between the actual chart area and the margins.}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
metricsgraphics object
}
\description{
\code{mjs_plot()} initializes the metricsgraphics.js html widget
and takes a data frame & variables holding character strings for
the  x & y column names as minimum input.
This must be used with or piped to a "geom" (metricsgraphics.js only supports single
"geom" layers) and can also be piped to other \code{mjs_} functions that
manipulate aesthetics.
}
\details{
See \href{http://metricsgraphicsjs.org/}{MetricsGraphics.js} for more information.
}

