% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_point}
\alias{mjs_point}
\title{metricsgraphics.js scatterplot "geom"}
\usage{
mjs_point(mjs, least_squares = FALSE, size_accessor = NULL,
  color_accessor = NULL, x_rug = FALSE, y_rug = FALSE)
}
\arguments{
\item{mjs}{plot object}

\item{least_squares}{add a least squares line? (default: \code{FALSE} - no)}

\item{size_accessor}{bare name of a column to use to scale the size of the points}

\item{color_accessor}{bare name of a column to use to scale the color of the points}

\item{x_rug}{show a "rug" plot next to the x axis? (default: \code{FALSE} - no)}

\item{y_rug}{show a "rug" plot next to the y axis? (default: \code{FALSE} - no)}
}
\description{
This function adds a point/scatterplot "geom" to a metricsgraphics.js html widget.
}
\examples{
\dontrun{
mtcars \%>\%
 mjs_plot(x=wt, y=mpg, width=400, height=300) \%>\%
 mjs_point(least_squares=TRUE)
}
}

