% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{renderMetricsgraphics}
\alias{renderMetricsgraphics}
\title{Widget render function for use in Shiny}
\usage{
renderMetricsgraphics(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

