% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metricsgraphics-package.R
\name{metricsgraphics-exports}
\alias{JS}
\alias{\%>\%}
\alias{metricsgraphics-exports}
\alias{saveWidget}
\title{metricsgraphics exported operators}
\description{
The following functions are imported and then re-exported
from the dygraphs package to enable use of the magrittr
pipe operator with no additional library calls
}

