% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metricsgraphics.r
\name{metricsgraphicsOutput}
\alias{metricsgraphicsOutput}
\title{Widget output function for use in Shiny}
\usage{
metricsgraphicsOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output id}

\item{width}{width}

\item{height}{height}
}
\description{
Widget output function for use in Shiny
}

