% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metricsgraphics.R
\name{mjs_axis_x}
\alias{mjs_axis_x}
\title{Configure x axis ticks & limits}
\usage{
mjs_axis_x(mjs, show = TRUE, xax_count = 6, min_x = NULL, max_x = NULL,
  extended_ticks = FALSE, xax_format = "plain",
  show_secondary_x_label = NULL, rug = FALSE)
}
\arguments{
\item{mjs}{plot object}

\item{show}{display the axis? (default: \code{TRUE} - yes)}

\item{xax_count}{tick count}

\item{min_x}{min limit for x axis}

\item{max_x}{max limit for x axis}

\item{extended_ticks}{extend ticks on x axis?}

\item{xax_format}{how to format tick labels. Currently one of "plain", "comma"
or "date"}

\item{show_secondary_x_label}{determines whether to show the year, or another
unit of time in the case of smaller series, on the x-axis below
the x-axis labels.}

\item{rug}{show a "rug" plot next to the x axis? (default: \code{FALSE} - no)}
}
\description{
Configure x axis ticks & limits
}
\note{
xax_format is likely to undergo a drastic change in future releases but
      support for these three formats will also likely remain.
}

