% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis.r
\name{mjs_axis_y}
\alias{mjs_axis_y}
\title{Configure y axis ticks & limits}
\usage{
mjs_axis_y(mjs, show = TRUE, yax_count = 5, min_y = NULL, max_y = NULL,
  extended_ticks = FALSE, y_scale_type = "linear", yax_units = "",
  rug = FALSE)
}
\arguments{
\item{mjs}{metricsgraphics plot object}

\item{show}{display the axis? (default: \code{TRUE} - yes)}

\item{yax_count}{tick count}

\item{min_y}{min limit for y axis}

\item{max_y}{max limit for y axis}

\item{extended_ticks}{extend ticks on y axis?}

\item{y_scale_type}{scale for y axis; either "linear" (default) or "log"}

\item{yax_units}{a prefix symbol to be shown alongside the y axis' labels.
Useful for currencies, for instance.}

\item{rug}{show a "rug" plot next to the y axis? (default: \code{FALSE} - no)}
}
\value{
metricsgraphics object
}
\description{
Configure y axis ticks & limits
}

