% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.r
\name{mjs_labs}
\alias{mjs_labs}
\title{Configure axis labels & plot description}
\usage{
mjs_labs(mjs, x_label = NULL, y_label = NULL)
}
\arguments{
\item{mjs}{metricsgraphics object}

\item{x_label}{label for x asis}

\item{y_label}{label for y axis}
}
\value{
metricsgraphics object
}
\description{
Configure axis labels & plot description
}
\examples{
mtcars \%>\%
 mjs_plot(x=wt, y=mpg, width=400, height=300) \%>\%
 mjs_point(color_accessor=carb, size_accessor=carb) \%>\%
 mjs_labs(x="Weight of Car", y="Miles per Gallon")

}

